+++ changelog.md
@@ -10,6 +10,8 @@ Trivial changes when a new feature is added are not denoted.
     - In Explore variable pairs
     - in Compare variables
     - in Compare groups
+- Display results in groups of raw data, sample properties and population properties
+    - in Compare groups 
 - Rank information is displayed for ordinal variables:
     - In single variable
     - In variable pairs (in scatterplot)